Populate MessageDetails.SignatureError when unverified
It's invalid for library users to look at MessageDetails.SignatureError
before the OpenPGP message is fully read. Populate it with an error
(cleared once the message is fully read).

This is technically a breaking change, but code looking at
SignatureError before the full message is read is broken anyways.

References: https://github.com/ProtonMail/go-crypto/issues/92