Add support to only perform an s2k key derivation once while encrypting/decrypting multiple private keys
In previous versions, it is not possible to encrypt multiple private keys with the
same encryption key derived from a passphrase.
A fresh encryption key is derived for each encryption and
similar s2k key derivations are repeated for decryption.
This commit adds support for encrypting/decrypting multiple private keys with a single encryption
key derived with a single s2k key-derivation:

- Adds two high-level methods on `Entity` called  'LockAllKeys' and 'UnlockAllKeys'

- Adds two functions  in `packet` called `PrivateKeysEncrypt`  and ` PrivateKeysDecrypt`
that allow encrypting/decrypting multiple `PrivateKeys.`

- Refactoring in PrivateKey encryption and decryption

- Adds  `Cache` in the `s2k` package to avoid recomputing similar
 key derivations in `packet.PrivateKeysDecrypt`

- Changes the type of the `s2k.Params` salt field to byte array
 and introduces a getter for retrieving the salt.