Add support for v6 signatures
This change implements the necessary updates for parsing and serializing v6 signature
packets. Further it implements the v6 signature salt. V6 signatures require that a salt is
prepended to the signing content before creating the signature hash. In this commit, we
add this feature, which requires a change in the API for signing content.
The Sign API takes an externally created hasher object as an argument to sign.
To ensure that the salt is hashed before the conten in v6 singatures, we introduce a
PrepareSignature method that creates a hasher object with prepended salt.