Add methods on Entity to encrypt/decrypt all private keys (#156)
This change introduces two new functions, Entity.EncryptPrivateKeys and
DecryptPrivateKeys, which respectively encrypt and decrypt all
non-dummy private keys in the entity with the same passphrase.

On encryption, the derived key is computed once and reused for every
key; while on decryption, derived keys are cached, such that if the
same S2K parameters were used during encryption for all keys, the
derived key is only computed once.